Remove deprecated `useAIFeatureEnabled` hook and update imports across components. Refactor to utilize the new hook from the user state module for AI feature management.
