+++ packages/commonwealth/client/scripts/state/ui/user/index.ts
@@ -1,4 +1,5 @@
 import { useLocalAISettingsStore } from './localAISettings';
+import { useAIFeatureEnabled } from './useAIFeatureEnabled';
 import { useUserStore, userStore } from './user';
-export { useLocalAISettingsStore, userStore };
+export { useAIFeatureEnabled, useLocalAISettingsStore, userStore };
 export default useUserStore;
