+++ packages/commonwealth/client/scripts/state/ui/user/useAIFeatureEnabled.ts
