+++ packages/commonwealth/client/scripts/views/components/Comments/CommentEditor/CommentEditor.tsx
@@ -2,7 +2,6 @@ import { ContentType } from '@hicommonwealth/shared';
 import clsx from 'clsx';
 import { notifyError } from 'controllers/app/notifications';
 import { isCommandClick } from 'helpers';
-import { useAIFeatureEnabled } from 'hooks/useAIFeatureEnabled';
 import Account from 'models/Account';
 import Thread from 'models/Thread';
 import type { DeltaStatic } from 'quill';
@@ -10,7 +9,7 @@ import React, { forwardRef, useCallback, useEffect, useState } from 'react';
 import { useAiCompletion } from 'state/api/ai';
 import { generateCommentPrompt } from 'state/api/ai/prompts';
 import useGetCommunityByIdQuery from 'state/api/communities/getCommuityById';
-import { useLocalAISettingsStore } from 'state/ui/user';
+import { useAIFeatureEnabled, useLocalAISettingsStore } from 'state/ui/user';
 import { useTurnstile } from 'views/components/useTurnstile';
 import { User } from 'views/components/user/user';
 import { jumpHighlightComment } from 'views/pages/discussions/CommentTree/helpers';
