+++ packages/commonwealth/client/scripts/views/components/StickEditorContainer/StickyInput/StickyInput.tsx
@@ -1,7 +1,6 @@
 import { CompletionModel, ContentType } from '@hicommonwealth/shared';
 import ClickAwayListener from '@mui/base/ClickAwayListener';
 import { notifyError } from 'controllers/app/notifications';
-import { useAIFeatureEnabled } from 'hooks/useAIFeatureEnabled';
 import useBrowserWindow from 'hooks/useBrowserWindow';
 import { Thread } from 'models/Thread';
 import type { Topic } from 'models/Topic';
@@ -19,7 +18,7 @@ import {
   generateThreadPrompt,
 } from 'state/api/ai/prompts';
 import useSidebarStore from 'state/ui/sidebar';
-import { useLocalAISettingsStore } from 'state/ui/user';
+import { useAIFeatureEnabled, useLocalAISettingsStore } from 'state/ui/user';
 import { AIModelSelector } from 'views/components/AIModelSelector';
 import type { CommentEditorProps } from 'views/components/Comments/CommentEditor/CommentEditor';
 import CommentEditor from 'views/components/Comments/CommentEditor/CommentEditor';
